fix/move-widget-class
