+++ classes/widgets/class-badge-streak.php
@@ -7,8 +7,6 @@
 
 namespace Progress_Planner\Widgets;
 
-use Progress_Planner\Widget;
-
 /**
  * Badge_Streak class.
  */
