+++ classes/widgets/class-latest-badge.php
@@ -7,8 +7,6 @@
 
 namespace Progress_Planner\Widgets;
 
-use Progress_Planner\Widget;
-
 /**
  * Latest_Badge class.
  */
