+++ classes/widgets/class-published-content.php
@@ -7,8 +7,6 @@
 
 namespace Progress_Planner\Widgets;
 
-use Progress_Planner\Widget;
-
 /**
  * Published_Content class.
  */
